backend/requirements.txt
flask
gunicorn
groq
python-dotenv
requests
google-cloud-dialogflow
pandas